tab bar widget seperated
